requirements.txt
flake8-debugger==3.0.0
flake8-isort==2.3
flake8-print==3.0.1
flake8==3.5.0
isort==4.2.15
pytest==3.3.1
python-dotenv==0.7.1
